// src/types/pokemon.ts
export interface Pokemon {
    abilities: Ability[]
    base_experience: number
    cries: Cries
    forms: Form[]
    game_indices: Index[]
    height: number
    held_items: any[]
    id: number
    is_default: boolean
    location_area_encounters: string
    moves: Mfe[]
    name: string
    order: number
    past_abilities: any[]
    past_types: any[]
    species: Species
    sprites: Sprites
    stats: Stat[]
    types: Type[]
    weight: number
}


interface Ability {
    ability: Ability2
    is_hidden: boolean
    slot: number
}

interface Ability2 {
    name: string
    url: string
}

interface Cries {
    latest: string
    legacy: string
}

interface Form {
    name: string
    url: string
}

interface Index {
    game_index: number
    version: Version
}

interface Version {
    name: string
    url: string
}

interface Mfe {
    move: Move
    version_group_details: VersionGroupDetail[]
}

interface Move {
    name: string
    url: string
}

interface VersionGroupDetail {
    level_learned_at: number
    move_learn_method: MoveLearnMethod
    version_group: VersionGroup
}

interface MoveLearnMethod {
    name: string
    url: string
}

interface VersionGroup {
    name: string
    url: string
}

interface Species {
    name: string
    url: string
}

interface Sprites {
    back_default: string
    back_female: any
    back_shiny: string
    back_shiny_female: any
    front_default: string
    front_female: any
    front_shiny: string
    front_shiny_female: any
    other: Other
    versions: Versions
}

interface Other {
    dream_world: DreamWorld
    home: Home
    "official-artwork": OfficialArtwork
    showdown: Showdown
}
// ...
interface Versions {
    "generation-i": GenerationI
    "generation-ii": GenerationIi
    "generation-iii": GenerationIii
    "generation-iv": GenerationIv
    "generation-v": GenerationV
    "generation-vi": GenerationVi
    "generation-vii": GenerationVii
    "generation-viii": GenerationViii
}
// ...
interface Stat {
    base_stat: number
    effort: number
    stat: Stat2
}

interface Stat2 {
    name: string
    url: string
}

interface Type {
    slot: number
    type: Type2
}

interface Type2 {
    name: string
    url: string
}